Fix the CRY installation
Ratpac requires the CRYLIB, CRYINCLUDE and CRYDATA variables to be set in order to add CRY to the ratpac install. These variables weren't being exported before the ratpac install, hence CRY wasn't being added to ratpac. I've added the lines to export these variables before ratpac is built, but only if libCRY.so exists in the ratpac-setup lib folder.