Add installation steps for xerces and icu and change shebang
Xerces and ICU seem to be common missing dependencies, so I've added them into the installation.

The sheband in setup.sh now uses '/usr/bin/env bash', which should get the first instance of bash in the PATH variable. This should avoid problems with multiple installations of bash (a particular problem with homebrew) and is a portable method across systems.