Fix issue on macos where install used high memory
The variable `procuse` that passes on the number of cores to use in the build was being set after options[procuse] was defined. Hence when `make -j"${options[procuse]}"` was being called, it resulted in `make -j` being executed. Not sure why this led to lots of memory usage at the build stage, but setting `procuse` at the correct point seems to have fixed things.

Also updated the docs with a mention of the MacOS installation.